wip
This is an in-progress commit for enabling method calls when
a variable type is a union of emitter types.

As long as all emitter types in the union have an event,
you'll be able to emit or add a listener with type safety.

In past versions of ee-ts (and also in strict-event-emitter-types),
this cannot be done, since every event is jammed into a single
method signature. To solve this, each event needs its own method
signature, so TypeScript can detect the compatible events.